backend/todo_api/requirements.txt
Django==4.2.17
djangorestframework==3.15.0
django-cors-headers==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==21.2.0
